small internal refactor
