+++ bidict/_abc.py
@@ -4,7 +4,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-"""Provides ABCs, DuplicationBehaviors, and BidictExceptions."""
+"""Provides bidict ABCs."""
 
 from abc import abstractproperty
 from collections import Mapping
@@ -54,67 +54,3 @@ class BidirectionalMapping(Mapping):
         if mro is None:
             return NotImplemented
         return all(any(B.__dict__.get(i) for B in mro) for i in cls._subclsattrs)
-
-
-class _Marker(object):
-    def __init__(self, name):
-        self.name = name
-
-    def __repr__(self):
-        return '<%s>' % self.name  # pragma: no cover
-
-
-class DuplicationBehavior(_Marker):
-    """
-    Provide RAISE, OVERWRITE, IGNORE, and ON_DUP_VAL duplication behaviors.
-
-    .. py:attribute:: RAISE
-
-        Raise an exception when a duplication is encountered.
-
-    .. py:attribute:: OVERWRITE
-
-        Overwrite an existing item when a duplication is encountered.
-
-    .. py:attribute:: IGNORE
-
-        Keep the existing item and ignore the new item when a duplication is
-        encountered.
-
-    .. py:attribute:: ON_DUP_VAL
-
-        Used with *on_dup_kv* to specify that it should match whatever the
-        duplication behavior of *on_dup_val* is.
-    """
-
-
-DuplicationBehavior.RAISE = RAISE = DuplicationBehavior('RAISE')
-DuplicationBehavior.OVERWRITE = OVERWRITE = DuplicationBehavior('OVERWRITE')
-DuplicationBehavior.IGNORE = IGNORE = DuplicationBehavior('IGNORE')
-DuplicationBehavior.ON_DUP_VAL = ON_DUP_VAL = DuplicationBehavior('ON_DUP_VAL')
-_MISS = _Marker('MISSING')
-
-
-class BidictException(Exception):
-    """Base class for bidict exceptions."""
-
-
-class DuplicationError(BidictException):
-    """Base class for exceptions raised when uniqueness is violated."""
-
-
-class KeyDuplicationError(DuplicationError):
-    """Raised when a given key is not unique."""
-
-
-class ValueDuplicationError(DuplicationError):
-    """Raised when a given value is not unique."""
-
-
-class KeyAndValueDuplicationError(KeyDuplicationError, ValueDuplicationError):
-    """
-    Raised when a given item's key and value are not unique.
-
-    That is, its key duplicates that of another item,
-    and its value duplicates that of a different other item.
-    """
