+++ bidict/_miss.py
@@ -1,0 +1,12 @@
+# Copyright 2017 Joshua Bronson. All Rights Reserved.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+
+"""Provides the "MISSING" marker."""
+
+from ._marker import _Marker
+
+
+_MISS = _Marker('MISSING')
